Refactor transaction input data handling by moving getTxInputData to a new file and updating imports accordingly
