+++ core/ui/mpc/keysign/action/mutations/useKeysignMutation.ts
@@ -13,8 +13,8 @@ import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignM
 import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { useKeysignAction } from '@core/ui/mpc/keysign/action/state/keysignAction'
 import { customMessageConfig } from '@core/ui/mpc/keysign/customMessage/config'
+import { getTxInputData } from '@core/ui/mpc/keysign/tx/inputData/getTxInputData'
 import { getKeysignChain } from '@core/ui/mpc/keysign/utils/getKeysignChain'
-import { getTxInputData } from '@core/ui/mpc/keysign/utils/getTxInputData'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { chainPromises } from '@lib/utils/promise/chainPromises'
